Import Upstream version 2.25

+++ Contrib/zip2exe/main.cpp
@@ -412,9 +412,9 @@ DWORD WINAPI ThreadProc(LPVOID p) // thread that will start & monitor wwwinamp
     return 1;
   }
 
-  unsigned long exit=0;  //process exit code
-  unsigned long bread;   //bytes read
-  unsigned long avail;   //bytes available
+  DWORD exit=0;  //process exit code
+  DWORD bread;   //bytes read
+  DWORD avail;   //bytes available
 
   memset(buf,0,sizeof(buf));
   while (1)      //main program loop
@@ -668,7 +668,7 @@ BOOL CALLBACK DlgProc(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM lParam)
     case WM_DROPFILES:
     {
       char dropped_file[MAX_PATH]="";
-      if (DragQueryFile((HDROP)wParam,-1,NULL,0)==1)
+      if (DragQueryFile((HDROP)wParam,(UINT)-1,NULL,0)==1)
       {
         DragQueryFile((HDROP)wParam,0,dropped_file,MAX_PATH);
         if (lstrlen(dropped_file)>0)
